cf1_get_file_upload_to_gcs/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage
wget